Use a default --disk argument for virt-install
In all our VMS we've been using:

    --disk vol=pool/image_name,bus=virtio,cache=none,io=native

What was a bit bothersome was that the image_name needed to be changed in
each definition file. With this commit we inject the above line with the
correct pool/image names into the "virt-install" command.

If further parametrization/flexibility is required in the future, we can
revisit this.